backend/requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
cloudinary==1.29.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.5
django-cors-headers==3.12.0
django-dotenv==1.4.2
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.7
MarkupSafe==2.1.1
packaging==21.3
pyparsing==3.0.9
pytz==2022.1
PyYAML==6.0
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.9
zipp==3.8.0
